status: Clear the initialized state more quickly
When the interval is long, it may be several minutes before the
operator reports progressing false. Make that happen more quickly
by checking the client enabled status up front.

Also allow the upload interval to be configured in the secret -
this will eventually be replaced by real config, but we want CI
to report every 10 minutes so we are guaranteed to get some
output.