auth: Invalid token is reported as error in cluster status
If the token for the cluster contains a newline, the error is
reported in a generic fashion to the cluster operator status which
contains the auth token value. This is an information leak and
should not happen.

Prevent future disclosure bugs by not printing the exact error when
we build an invalid request. Check the token value in both the client
(before setting the header) and when we read from config (before
loading the token) so that we report this error in an earlier place
with a specific error.

To reproduce, put a newline inside the cloud.openshift.com token in
the cluster pull-secret (in openshift-config namespace).