Bug 1745973 - don't degrade on upload error
Given the upload failures are outside of control of the OCP administrator
and they are not fatal for the the functionality of the cluster,
don't set the cluster as degraded for a single upload failures. There is
a treshold for how many upload failures we tolerate before starting
reporting the issue.

Separate state "UploadDegraded" is introduced to indicate the health of the
upload process.