gather: Add a workload fingerprint gatherer
The rough structure and shape of workloads on the cluster provides
information that is valuable for support scenarios as well as future
development. The size, scale, and scope of pods as well as use of
init containers is relevant for assessing how to prioritize
improvements to the project. As we see increasing amounts of supply
chain security scrutiny, the image hashes running on a cluster can
potentially be used to highly compromised images or to identify when
certain workloads or known images are highly out of date. Finally,
for a subset of workloads with known executables the info about the
command being executed (in a hashed version) would allow for analysis
of correlations between usage and failures for support.

This gatherer returns no directly identifiable workload information
unless a hash collision is feasible. Only namespace names and the
first values of command and argument are assessed for now, and command
and argument are reduced from their full values to a partial identifier
prior to hashing:

"-config.file=/etc/cache-config/cache-conf.yaml" to
"-config.file" to
"V4JjWhyBxoZx"

Images that have metadata in the cluster are associated with the
requests and a running cache is kept to minimize accesses. The
gatherer single threads image lookups at 10qps and will exit if the
total gather time exceeds a reasonable constant, with subsequent calls
potentially being faster due to caching.

The gatherer attempts to minimize the amount of info held in memory
and should be reasonably resource efficient. If necessary, the
gatherer could be executed at a less frequent interval.