Bug 1942271: Gather openshift-cluster-version pods and events (#381)
* pkg/insights/insightsclient: Inline client handling for ClusterVersion

Using the native client code directly, instead of through gather's
wrapper, is not much of a convenience hit.  And this gives the client
space to improve its client persistence going forward.  Completes the
transition begun in 73d3cfd8d1 (Move gather function execution into
the Gatherer, also removes the link between gathering and
insightsclient, 2020-12-03, #279) to decouple the client from the
gatherer.  This also allows me to make getClusterVersion internal on
the gather side, so I can shuffle it's API a bit in future commits
without breaking any consumers.

* pkg/gather/clusterconfig: Collapse GatherClusterID into GatherClusterVersion

There's no reason to fetch the ClusterVersion twice, even if we are
creating two Records based on its content.  This also sets the stage
for gathering additional items like sad cluster-version operator pods.

I personally don't see a need to call out the fact that ClusterVersion
included clusterID, but Tomas wanted it [1].

[1]: https://github.com/openshift/insights-operator/pull/381#discussion_r601169083

* pkg/gather/clusterconfig/version: Walk openshift-cluster-version

Using similar logic to what gatherClusterOperators uses, but for the
ClusterVersion operator's namespace.  Usually we'd only gather the pod
YAML if the pod was failing, but in order to help audit tolerations
for [1], at the moment I'm gathering it every time.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1941901

* docs/gathered-data: Regenerate

Generated with:

  $ make gen-doc