[release 4.6] Bug 1953582: GatherClusterOperators and GatherClusterOperatorsPodAndEvents (#413)
* Bug 1951034: Split up the GatherClusterOperators into smaller parts (#397)

* Split operators gather

This splits the cluster operators gather into two gathers, one for operator resources and other for unhealty operators.

* Adds precommit target

This adds the precommit target to the makefile, so we can execute it to test the stashed changes before actually commit.

* Move CompactedEvent

This move the structure that was previously defined on operators to operators_unhealthy.

* Change operators name

This changes the previous operator_unhealty to operator_pods_and_events.

* Update gathered documentation

This updates the gatherer documentation for operators and operators_pods_and_events

* Skip Test_UnhealtyOperators_FetchPodContainerLog

This removes and skips with note the Test_UnhealtyOperators_FetchPodContainerLog test