Adjust the Pod log gather
this limits the second log gather if found a possible stacktrace message, keep the log with an offset before the first stacktrace message line and keep the rest of the log intact. The normal log should be limited by logMasTailLines and the stacktrace log should not be larger than logMaxLongTailLines.