Better pod log gathering with offset for stacktrace messages (#456)
* handling pod log gathering with offset for stacktracing

* Change the regex rule

* Improving stacktrace log gathering

* Simplify stacktrace regex

* Adjust the Pod log gather

this limits the second log gather if found a possible stacktrace message, keep the log with an offset before the first stacktrace message line and keep the rest of the log intact. The normal log should be limited by logMasTailLines and the stacktrace log should not be larger than logMaxLongTailLines.

* Tunning the fetch log parameters

* logLinesOffset as int

* I'll not use pointer here

* Clean up the logic for getLogWithStacktracing

* Removes the circular buffer