DO NOT MERGE: Add a name analysis command
The `workloadinfo` command performs some transformation passes on
the config/workload_info.json file that makes it more amenable to
human exploration. The first implemented pass takes a dictionary of
names that represent the namespace, firstArg, or firstCommand fields
prior to hashing, creates an index between the hash and the source
name, and then replaces those inline into the workload info struct
which is serialized to stdout. The `--verbose` flag will print some
stats about how many names were identified. Any non-matching hashes
are left as is in the output.

A helper flag `--extract-names` was added to bulk transform exemplar
command lines (as might be found in image metadata or pod command and
args fields) into names for inlining.

Expose the necessary fields and types from pkg/gatherers/workloads
to avoid reuse.

This command would be used in the future to perform other
transformations for human use such as taking known maps of image
SHAs to metadata, or calculating other stats about usage from a
workload info file.