Collect oscp CRs
Control Plan of the Red Hat OpenStack starting from the version 18
(RHOSO 18) will be deployed using OpenShift operators as a set of
OpenShift applications.
Compute nodes will be deployed still as RHEL nodes and will be outside
of the OpenShift cluster but all will be mangaged by the
Dataplane-operator and only interface for OpenStack operators will be
though the CRs:

- OpenStackControlPlane CR for the control plane,
- OpenStackDataplaneNodeSet and OpenStackDataplaneDeployment CRs for the
  compute nodes.

This patch adds OpenStackControlPlane CR to the insights-operator so that
whole configuration of the OpenStack control plane may be gathered by the
insights-operator and send to analysis to the Insights server.

Co-Authored-By: Yatin Karel <ykarel@redhat.com>

Conflicts:
    docs/gathered-data.md
    pkg/gatherers/clusterconfig/clusterconfig_gatherer.go

Related: #OSPRH-5904
(cherry picked from commit 999cbf54ecc1dbe081d7c48565b713766e8151db)