Add helper function to anonymize custom fields in the OpenStack CRs
Some fields in the OpenStack related CRs may exist in the different
paths in the CR structure, for example 'ansibleUser' may be defined in
the 'spec.nodeTemplate' but also in 'spec.nodes.<node_name>' but in both
cases it should be anonymized.
This patch adds helper function which goes over whole
map[string]interface{} data structure, looks for the fields to anonymize
(by the field's name) and anonymize them.

Related: OSPRH-5904