OSPRH-5904: Integration of the OpenStack CRs into the insights-operator (#951)
* Collect oscp CRs

Control Plan of the Red Hat OpenStack starting from the version 18
(RHOSO 18) will be deployed using OpenShift operators as a set of
OpenShift applications.
Compute nodes will be deployed still as RHEL nodes and will be outside
of the OpenShift cluster but all will be mangaged by the
Dataplane-operator and only interface for OpenStack operators will be
though the CRs:

- OpenStackControlPlane CR for the control plane,
- OpenStackDataplaneNodeSet and OpenStackDataplaneDeployment CRs for the
  compute nodes.

This patch adds OpenStackControlPlane CR to the insights-operator so that
whole configuration of the OpenStack control plane may be gathered by the
insights-operator and send to analysis to the Insights server.

Co-Authored-By: Yatin Karel <ykarel@redhat.com>

Related: #OSPRH-5904

* Add anonymization for data in the oscp CR

Some fields in the collected OpenStackControlPlane CRs should be
anonymized. This is for example field 'options' from the 'dns'
service spec.

Related: #OSPRH-5904

* Add test module for the oscp CR gatherer

* Add way to drop some fields from the gathered oscp CR

For now it will drop only
'metadata/annotations/kubectl.kubernetes.io/last-applied-configuration'
but other fields can be added to that list too.

* Collect osdpns CRs

Control Plan of the Red Hat OpenStack starting from the version 18
(RHOSO 18) will be deployed using OpenShift operators as a set of
OpenShift applications.
Compute nodes will be deployed still as RHEL nodes and will be outside
of the OpenShift cluster but all will be mangaged by the
Dataplane-operator and only interface for OpenStack operators will be
though the CRs:

- OpenStackControlPlane CR for the control plane,
- OpenStackDataplaneNodeSet and OpenStackDataplaneDeployment CRs for the
  compute nodes.

This patch adds OpenStackDataplaneNodeSet CR to the insights-operator so that
whole configuration of the OpenStack control plane may be gathered by the
insights-operator and send to analysis to the Insights server.

Co-Authored-By: Yatin Karel <ykarel@redhat.com>

Related: #OSPRH-5904

* Collect osdpd CRs

Control Plan of the Red Hat OpenStack starting from the version 18
(RHOSO 18) will be deployed using OpenShift operators as a set of
OpenShift applications.
Compute nodes will be deployed still as RHEL nodes and will be outside
of the OpenShift cluster but all will be mangaged by the
Dataplane-operator and only interface for OpenStack operators will be
though the CRs:

- OpenStackControlPlane CR for the control plane,
- OpenStackDataplaneNodeSet and OpenStackDataplaneDeployment CRs for the
  compute nodes.

This patch adds OpenStackDataplaneDeployment CR to the insights-operator so that
whole configuration of the OpenStack control plane may be gathered by the
insights-operator and send to analysis to the Insights server.

Co-Authored-By: Yatin Karel <ykarel@redhat.com>

Related: #OSPRH-5904

* Add 'ansibleHost' field to be annonymized

* Rename names of the collected OpenStackDataPlane resources

* Update examples of the OpenStack CRs to their latest versions

Related: #OSPRH-5904

* Collect OpenStackVersion CRs

Red Hat's openstack-operator in addition to the OpenStackControlPlane
and OpenStackDataPlane CRs uses also additonal CR 'openstackversions'
which reports version of the openstack(s) deployed in the OCP cluster.
This patch adds support for gathering also this CRs from the cluster if
they exists.

Related: #OSPRH-5904