OCPBUGS-56780: Report an error when the insights-runtime-extractor is not running (#1088)
If there are no running pods for the insights-runtime-extractor, report it as an error.
There should be 1 running pods on every cluster's worker node. The number can differ based on the
cluster size but the absence of any running pods is an error to report to the user.

This fixes https://issues.redhat.com/browse/OCPBUGS-56780
Upstream issue: https://issues.redhat.com/browse/OCPBUGS-55804

Signed-off-by: Jeff Mesnil <jmesnil@redhat.com>