CCXDEV-15045: update code to use v1alpha2 api version (#1093)
* feat: update CRDs to v1alpha2

This commit is updating CRDs to a v1alpha2 versions
and removing the v1alpha1 resources.

Signed-off-by: Ondrej Pokorny <opokorny@redhat.com>

* feat: update api and client packages to v1alpha2

This commit is updating all the code that is using
the api and client packages from v1alpha1 to v1alpha2
version.

Signed-off-by: Ondrej Pokorny <opokorny@redhat.com>

* feat: rename dataPolicy to dataPolicies

This commit is renaming the dataPolicy field to
dataPolicies which reflects the current structure
better.

Signed-off-by: Ondrej Pokorny <opokorny@redhat.com>

---------

Signed-off-by: Ondrej Pokorny <opokorny@redhat.com>